millionx-phase2/snowflake/requirements.txt
# Snowflake Batch Sink Requirements

pandas==2.1.4
snowflake-connector-python[pandas]==3.5.0
kafka-python==2.0.2
python-dotenv==1.0.0
